add FlyWay. Tests. Error page. AccessDeniedHandler. New feature OnCloseFaculty. Prepare for show

+++ src/main/java/com/art/dip/utility/dto/CertificateDTO.java
@@ -7,7 +7,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class CertificateDTO {
 
 	@Range(min = 1,max = 10)
-	private Integer mark;
+	private Double mark;
 	
 	private MultipartFile file;
 }
